moved CuFile to cufile.py and added IOFuture
